Move DB instantiation and aquisition logic into DB modules
Previously, creating a SQLite DB file, creating read-only and read-write connection pools, and gatekeeping connection types (e.g., `ReadWrite`)
against the Kupo instance type (e.g., read only replica) was exposed by the `SQLite` module and performed in `Kupo`.

Now, those functions are encapsulated within the DB-specific modules to allow unique implementations. For instance, PostgreSQL does not have a DB file
and does not need separate pools for read-only and read-write connections.
